Remove legacy store.ts module, rename backend.ts to store.ts
This also removes all code for features which were still developed for the old server.ts, fixes #161
