+++ apps/frontend/src/main.ts
@@ -8,7 +8,7 @@ import 'roboto-fontface/css/roboto/roboto-fontface.css';
 import 'material-design-icons-iconfont/dist/material-design-icons.css';
 import VueAnalytics from 'vue-analytics';
 import 'material-design-icons-iconfont/dist/material-design-icons.css';
-import {BackendModule} from './store/backend';
+import {ServerModule} from './store/server';
 import axios from 'axios';
 import SetupToasted from '@/plugins/SetupToasted';
 import Toasted from 'vue-toasted';
@@ -41,9 +41,9 @@ new Vue({
       error => {
         // If there is no backend token then it is safe to assume this request
         // originated from the login page and should not perform the logout action.
-        if (BackendModule.token !== '' && error.response.status === 401) {
+        if (ServerModule.token !== '' && error.response.status === 401) {
           // if we catch a 401 error
-          BackendModule.Logout();
+          ServerModule.Logout();
         }
         return Promise.reject(error); // reject the Promise, with the error as the reason
       }
