+++ apps/frontend/src/mixins/ServerMixin.ts
@@ -1,10 +1,10 @@
-import {BackendModule} from '@/store/backend';
+import {ServerModule} from '@/store/server';
 import {Vue, Component} from 'vue-property-decorator';
 
 @Component
 export default class ServerMixin extends Vue {
   //checks if heimdall is in server mode
   get serverMode() {
-    return BackendModule.serverMode;
+    return ServerModule.serverMode;
   }
 }
