+++ apps/frontend/src/store/store.ts
@@ -8,8 +8,7 @@ import {HashLookup} from '@/store/lookup_hashes';
 import {InspecIntake} from '@/store/report_intake';
 import {ColorHack} from '@/store/color_hack';
 import {AppInfo} from '@/store/app_info';
-import {Server} from './server';
-import {IBackendState} from './backend';
+import {IServerState} from './server';
 
 Vue.use(Vuex);
 
@@ -35,8 +34,7 @@ export interface StoreType {
   intake: InspecIntake;
   colors: ColorHack;
   info: AppInfo;
-  heimdallServer: Server;
-  backend: IBackendState;
+  server: IServerState;
 }
 const store = new Vuex.Store<StoreType>({});
